Beginning of Python backend
This is a small but significant piece of code for the Python background. I have been doing research on creating a backend while Chris makes a Javascript frontend. 

This code is meant to show how a Python script can be run on a local host, which should help the project. I will get rid of the "Hello World" and comments as we progress, but this is to show how a Python backend can work on a local host.